primera implementacion variable independiente v_factor, sin probar si funciona. Pequeña modificacion2, se me dan bien los indices estos XD

+++ test/two_b_problem.h
@@ -13,6 +13,6 @@ typedef struct {
 
 double *dist_verlet_2_body(cuerpo2d *planetas_t0, int frames, double dt);
 cuerpo2d *condiciones_iniciales_2_body(double r_orbita, double v_factor);
-test_result *test_two_b_p_v_factor(double *v_factors, double v_factors_len);
+test_result *test_two_b_p_v_factor(double *v_factors, int v_factors_len);
 
 #endif
